+++ src/lang/en/seat.php
@@ -1,8 +1,11 @@
 <?php
 
 return [
-    'client_id'       => 'Client ID',
-    'client_secret'   => 'Client Secret',
-    'bot_token'       => 'Bot Token',
-    'use_email_scope' => 'Use email as Unique ID? Email scope will be requested from users.',
+    'client_id'        => 'Client ID',
+    'client_secret'    => 'Client Secret',
+    'bot_token'        => 'Bot Token',
+    'use_email_scope'  => 'Use email as Unique ID? Email scope will be requested from users.',
+    'visible_roles'    => 'Manage visibleRoles for Discord Role filter',
+    'can_add_roles'    => 'Manage canAddRoles for Discord Role filter',
+    'can_remove_roles' => 'Manage canRemoveRoles for Discord Role filter',
 ];